+++ metadata.rb
@@ -35,13 +35,13 @@ attribute "glassfish/base_dir",
   :display_name => "GlassFish Base Directory",
   :description => "The base directory of the GlassFish install",
   :type => "string",
-  :default => "/usr/local/glassfish3"
+  :default => "/usr/local/glassfish-3.1.2"
 
 attribute "glassfish/domains_dir",
   :display_name => "GlassFish Domain Directory",
   :description => "The directory containing all the domain definitions",
   :type => "string",
-  :default => "/usr/local/glassfish3/glassfish/domains"
+  :default => "/usr/local/glassfish-3.1.2/glassfish/domains"
 
 attribute "glassfish/domains",
   :display_name => "GlassFish Domain Definitions",